Only block Livewire watchers for entangled data that triggered the request
